Implement and fix various issues with blog post router

+++ back-end/src/routers/blog-post-router.ts
@@ -1,26 +1,46 @@
 import { Router } from 'express'
-import { OK } from 'zod'
 import { pool } from '../helpers/pg-pool'
 import { generateBlogPostFetchQuery } from '../helpers/post-fetch-query-generators/generate-blog-post-fetch-query'
-import { CREATED } from '../helpers/status-codes'
-import { authOptional } from '../middleware/auth'
-import { BlogPostAddValidator, BlogPostFetchData, BlogPostFetchValidator } from '../validators/blog-post-validator'
+import { CREATED, FORBIDDEN, OK, RESOURCE_NOT_FOUND } from '../helpers/status-codes'
+import { authMandatory, authOptional } from '../middleware/auth'
+import { BlogPostAddValidator, BlogPostFetchData, BlogPostFetchValidator } from '../validators/blog-post-validators'
 
 const BlogPostRouter = Router()
 
-BlogPostRouter.post('/', (req, res) => {
-    const { posterId, contents, replyTo } = BlogPostAddValidator.parse(req.body)
-    const sql = 'INSERT INTO blog_post (poster_id, contents, reply_to) VALUES $1, $2, $3'
+BlogPostRouter.post('/', authMandatory, (req, res) => {
+    const { contents, replyTo } = BlogPostAddValidator.parse(req.body)
 
-    pool.query(sql, [posterId, contents, replyTo])
-        .then(() => res.status(201).json(CREATED))
+    const replyToVerifySql = 'SELECT id FROM blog_post WHERE id = $1'
+    pool.query(replyToVerifySql, [replyTo])
+        .then(result => {
+            if (replyTo !== undefined && result.rowCount === 0) res.status(404).json(RESOURCE_NOT_FOUND)
+            else {
+                const insertSql = 'INSERT INTO blog_post (poster_id, contents, reply_to) VALUES ($1, $2, $3)'
+
+                pool.query(insertSql, [req.body.id, contents, replyTo])
+                    .then(() => res.status(201).json(CREATED))
+            }
+        })
+
+    
 })
 
-BlogPostRouter.delete('/:id', (req, res) => {
-    const sql = 'DELETE FROM blog_post WHERE id = $1'
+BlogPostRouter.delete('/:id', authMandatory, (req, res) => {
+    const idVerifySql = 'SELECT poster_id FROM blog_post WHERE id = $1'
 
-    pool.query(sql, [req.params.id])
-        .then(() => res.json(200).json(OK))
+    pool.query(idVerifySql, [req.params.id])
+        .then(result => {
+            if (result.rowCount === 0) res.status(404).json(RESOURCE_NOT_FOUND)
+            else if (result.rows[0].poster_id !== req.body.id) res.status(403).json(FORBIDDEN)
+            else {
+                const deleteSql = 'DELETE FROM blog_post WHERE id = $1'
+
+                pool.query(deleteSql, [req.params.id])
+                    .then(() => res.status(200).json(OK))
+            }
+        })
+
+    
 })
 
 BlogPostRouter.get('/', authOptional, (req, res) => {
@@ -35,4 +55,8 @@ BlogPostRouter.get('/:id(\\d+)', (req, res) => {
     const sql = 'SELECT * FROM blog_post WHERE id = $1'
 
     pool.query(sql, [req.params.id])
+        .then(result => res.status(200).json(result.rows))
 })
+
+export { BlogPostRouter }
+
